refactor: dividir responsabilidades de schemas
