+++ backend/schemas/auth_schemas.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+from typing import Optional
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+class TokenData(BaseModel):
+    nombre_usuario: Optional[str] = None
